CommandLineParser to main
