+++ main/cmd/commands_test.go
@@ -4,8 +4,6 @@ import (
 	"bytes"
 	"flag"
 	"fmt"
-	"os"
-	"syscall"
 
 	"github.com/kkrull/gohttp/mock"
 	. "github.com/onsi/ginkgo"
@@ -14,91 +12,11 @@ import (
 	"github.com/kkrull/gohttp/main/cmd"
 )
 
-var _ = Describe("CliCommandParser", func() {
-	Describe("#Build", func() {
-		var (
-			parser     *cmd.CliCommandParser
-			interrupts chan os.Signal
-
-			command cmd.CliCommand
-			stderr  *bytes.Buffer
-		)
-
-		BeforeEach(func() {
-			interrupts = make(chan os.Signal, 1)
-			parser = cmd.NewCliCommandParser(interrupts)
-			stderr = &bytes.Buffer{}
-		})
-
-		Context("given --help", func() {
-			BeforeEach(func() {
-				command = parser.Parse([]string{"/path/to/gohttp", "--help"})
-			})
-
-			It("returns HelpCommand", func() {
-				Expect(command).To(BeAssignableToTypeOf(cmd.HelpCommand{}))
-			})
-			It("the HelpCommand is configured for the name of the program in the first argument", func() {
-				command.Run(stderr)
-				Expect(stderr.String()).To(HavePrefix("Usage of /path/to/gohttp"))
-			})
-			It("the HelpCommand shows usage for the gohttp arguments", func() {
-				command.Run(stderr)
-				Expect(stderr.String()).To(ContainSubstring("The root content directory"))
-				Expect(stderr.String()).To(ContainSubstring("The TCP port on which to listen"))
-			})
-		})
-
-		Context("given a complete configuration for the HTTP server", func() {
-			It("returns a RunServerCommand", func() {
-				parser = cmd.NewCliCommandParser(interrupts)
-				command = parser.Parse([]string{"gohttp", "-p", "4242", "-d", "/tmp"})
-				Expect(command).To(BeAssignableToTypeOf(cmd.RunServerCommand{}))
-			})
-
-			It("the command waits until a signal is sent to the interrupt signal channel", func() {
-				quitHttpServer := make(chan bool, 1)
-				parser = &cmd.CliCommandParser{
-					Interrupts: interrupts,
-					NewCommandToRunHTTPServer: func(string, string, uint16) (cmd.CliCommand, chan bool) {
-						return nil, quitHttpServer
-					}}
-
-				command = parser.Parse([]string{"gohttp", "-p", "4242", "-d", "/tmp"})
-				interrupts <- syscall.SIGINT
-				Eventually(quitHttpServer).Should(Receive())
-			})
-		})
-
-		Context("given unrecognized arguments", func() {
-			It("returns ErrorCommand", func() {
-				command = parser.Parse([]string{"gohttp", "--bogus"})
-				Expect(command).To(BeAssignableToTypeOf(cmd.ErrorCommand{}))
-			})
-		})
-
-		Context("when the path is missing", func() {
-			It("returns ErrorCommand", func() {
-				command = parser.Parse([]string{"gohttp", "-p", "4242"})
-				Expect(command).To(BeAssignableToTypeOf(cmd.ErrorCommand{}))
-			})
-		})
-
-		Context("when the port is missing", func() {
-			It("returns ErrorCommand", func() {
-				command = parser.Parse([]string{"gohttp", "-d", "/tmp"})
-				Expect(command).To(BeAssignableToTypeOf(cmd.ErrorCommand{}))
-			})
-		})
-	})
-})
-
 var _ = Describe("CliCommands", func() {
 	var (
-		command cmd.CliCommand
-		stderr  *bytes.Buffer
-		code    int
-		err     error
+		stderr *bytes.Buffer
+		code   int
+		err    error
 	)
 
 	BeforeEach(func() {
@@ -109,7 +27,7 @@ var _ = Describe("CliCommands", func() {
 		Describe("#Run", func() {
 			It("returns the error and an exit code of 1, to indicate failure", func() {
 				givenError := fmt.Errorf("bang")
-				command = cmd.ErrorCommand{Error: givenError}
+				command := cmd.ErrorCommand{Error: givenError}
 				code, err = command.Run(stderr)
 				Expect(code).To(Equal(1))
 				Expect(err).To(Equal(givenError))
@@ -120,7 +38,7 @@ var _ = Describe("CliCommands", func() {
 	Describe("HelpCommand", func() {
 		Describe("#Run", func() {
 			BeforeEach(func() {
-				command = cmd.HelpCommand{FlagSet: flag.NewFlagSet("widget", flag.ContinueOnError)}
+				command := cmd.HelpCommand{FlagSet: flag.NewFlagSet("widget", flag.ContinueOnError)}
 				code, err = command.Run(stderr)
 			})
 
@@ -136,8 +54,9 @@ var _ = Describe("CliCommands", func() {
 
 	Describe("RunServerCommand", func() {
 		var (
-			server mock.Server
-			quit   chan bool
+			command cmd.RunServerCommand
+			server  mock.Server
+			quit    chan bool
 		)
 
 		Describe("#Run", func() {
